POLCOMS style radiatve transfer for radiative fluxes
TRA/tradwl.F90 added. step.F90 and step_oce.F90 edited. 
Namelist added: namtra_dwl, with two keys:  ln_tradwl, ln_vary_lambda

This version is still in progress, as the variable hbatt from CO6 is no in CO8, and I need to confirm my approach is
valid.

This was crashing due to rlambda2 being infinity, when depth was 0. I have set a max value of 0.25, which is
equivalent to about 10cm of water. 


The version is committed in restrospect, so may not include all the changes to work.


git-svn-id: svn+ssh://forge.ipsl.jussieu.fr/ipsl/forge/projets/nemo/svn/NEMO/branches/UKMO/NEMO_4.0.4_CO9_shelf_climate@15313 4aad9cc9-4d31-0410-b1e8-ee312aa4b1ec