TRA/tradwl.F90 BUG FIX
The over-stratification problem I was having was due to a bug in the TRA/tradwl.F90 code.

Enda looked through my namelist, and made some suggestsion which could be causing the error. I ran a series of tests to see if any caused the error (none did).

By comparing output.namelist.dyn between my run and Enda's, I found some additional differences, so I ran some additional test.

I found namelist:namtrd ln_tra_trd was causing the problem. Setting it to .false. fixed it.

Somewhere, ln_tra_trd = .true. was causing the problem.

ln_tra_trd is defined in trd_oce.F90
read in from the name list in trdini.F90, and is queried in trdtra, where it is used to call trd_tra_iom.
The only other place its used in (my version of) nemo is in trdini.F90, where it sets l_trdtra = True... 

l_trdtra is all over NEMO, in 20 F90 files.

However, I noted l_trdtra is also in tradwl.F90 that I implemented. Looking l_trdtra in tradwl, it looked suspicious. The call to the trends functions had been commented out by CEOD, and it used pointers (pointing to ua and va as temp space??) to store changes to T and S. So as well as not reporting the changes, they could be overwriting the ua and va fields (the "after" 3d u and v velocity).

I have commented out these sections of code. At a later date, it may be useful to add them back in, in a way that is consistent with NEMO4, but it has not been done here.





--


git-svn-id: svn+ssh://forge.ipsl.jussieu.fr/ipsl/forge/projets/nemo/svn/NEMO/branches/UKMO/NEMO_4.0.4_CO9_shelf_climate@15634 4aad9cc9-4d31-0410-b1e8-ee312aa4b1ec